Begin player piece storing method change
Currently storing pieces in a vector. This wont be efficient later on as
the game will need to access pieces at a coordinate many times. A better
choice is to use an unordered_map. Added unordered_map to store player
pieces and implemented custom hash function for coordinate key copied from
cppreference.com on std::hash