Change gamemanager player sides to pointers
Changing to a pointer as the player destructor was called right after
the constructor was called, causing the pieces to be empty, in turn
causing a later segfault. Would need to do some research on if theres a
way to not do this as im unsure if the player class is best suited for
being on the stack or heap.